+++ xnxx_api/__init__.py
@@ -1,10 +1,10 @@
 # xnxx_api/__init__.py
 
 __all__ = [
-    "Client", "Core", "Quality", "Video", "Callback", "threaded", "default", "FFMPEG",
+    "Client", "BaseCore", "Video", "Callback",
     "errors", "consts", "search_filters", "category"
 ]
 
 # Public API from xnxx_api.py
-from xnxx_api.xnxx_api import Client, Core, Quality, Video, Callback, threaded, default, FFMPEG
+from xnxx_api.xnxx_api import Client, BaseCore, Video, Callback
 from xnxx_api.modules import errors, consts, category, search_filters